% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{fitLinearModel}
\alias{fitLinearModel}
\title{Linear Regression Function}
\usage{
fitLinearModel(formula, data)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{data}{A data.frame that contains the variables in the model.}
}
\value{
A list containing model coefficients, standard errors, t-statistics, p-values,
sigma squared, R-squared, adjusted R-squared, F-statistic, p-value of the F-statistic,
number of observations, and number of predictors.
}
\description{
Performs linear regression on the provided dataset using the ordinary least squares method.
The function calculates coefficients, residuals, various statistics including R-squared,
adjusted R-squared, and F-statistics, and returns a comprehensive summary of the model.
}
\examples{
data(iris)
model = fitLinearModel(Petal.Length ~ Petal.Width + Sepal.Length, iris)
print(model)

}
