% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show.R
\name{generate_data_and_fit}
\alias{generate_data_and_fit}
\title{Generate Random Data and Fit a Simple Linear Regression Model}
\usage{
generate_data_and_fit(n = 100)
}
\arguments{
\item{n}{An integer representing the number of data points to generate. Default is 100.}
}
\description{
This function generates random data that follows a linear relationship with added noise.
It then fits a simple linear regression model to the generated data and plots the actual
data points along with the fitted regression line.
}
\details{
The function generates random data for the independent variable \( x \) using a normal distribution.
The dependent variable \( y \) is created using the equation \( y = 2x + 3 + \epsilon \), where \( \epsilon \)
is random noise with standard deviation 0.5. After generating the data, the function fits a simple linear regression
model using the \code{simple_lm} function and then visualizes the results using the \code{plot_lm} function.
}
\examples{
# Generate random data and fit a linear model
generate_data_and_fit(n = 100)

}
\seealso{
\code{\link{simple_lm}} for fitting a simple linear regression model.

\code{\link{plot_lm}} for plotting the data and fitted regression line.
}
