% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{getFStatistic}
\alias{getFStatistic}
\title{Return the F statistics value for the model}
\usage{
getFStatistic(model)
}
\arguments{
\item{model}{The model list object returned by `fitLinearModel`.}
}
\value{
Returns F statistics for model
}
\description{
Return the F statistics value for the model
}
\examples{
data(iris)
model <- fitLinearModel(Petal.Length ~ Petal.Width + Sepal.Length, iris)
F_statistic <- getFStatistic(model)
print(F_statistic)

}
