% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{getRSquared}
\alias{getRSquared}
\title{Return the R Squared value for the model}
\usage{
getRSquared(model)
}
\arguments{
\item{model}{The model list object returned by `fitLinearModel`.}
}
\value{
Returns R Squared for model
}
\description{
Return the R Squared value for the model
}
\examples{
data(iris)
model <- fitLinearModel(Petal.Length ~ Petal.Width + Sepal.Length, iris)
R_squared <- getRSquared(model)
print(R_squared)

}
