% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{model_summary}
\alias{model_summary}
\title{Model Summary}
\usage{
model_summary(model)
}
\arguments{
\item{model}{The model list object returned by `fitLinearModel()`.}
}
\value{
Prints the summary table
}
\description{
The purpose of this function is to show a full summary of the linear regression model, including
estimates of the coefficients, standard errors, t-values, and p-values of fitted coefficients.
}
\examples{
data(iris)
model = fitLinearModel(Petal.Length ~ Petal.Width + Sepal.Length, iris)
model_summary(model)

}
