% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cdf.R
\name{plot_cdf}
\alias{plot_cdf}
\title{Plot Cumulative Distribution Function (CDF)}
\usage{
plot_cdf(x)
}
\arguments{
\item{x}{A numeric vector of data for which the cumulative distribution function will be calculated.}
}
\description{
This function plots the cumulative distribution function (CDF) of a given numeric vector.
It sorts the input data and calculates the cumulative probability, then plots the CDF.
}
\details{
The function first sorts the data in ascending order and calculates the cumulative probability
for each data point by dividing its rank by the total number of data points. The resulting CDF
is then plotted as a line graph with the x-axis representing the sorted data points and the y-axis
representing the cumulative probability.
}
\examples{
# Example data
x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)

# Plot the CDF for the data
plot_cdf(x)

}
\seealso{
\code{\link{hist}} for plotting a histogram.
}
