% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSE_Predict.R
\name{predict_least_squares}
\alias{predict_least_squares}
\title{Predict using the Least Squares Model}
\usage{
predict_least_squares(model, new_X)
}
\arguments{
\item{model}{A list returned by the \code{least_squares} function.}

\item{new_X}{A matrix or data frame of new predictor variables (independent variables). The first column should not include a column of ones.}
}
\value{
A numeric vector of predicted values.
}
\description{
This function predicts new response values using a fitted least squares model.
}
\examples{
X <- matrix(c(1, 2, 3, 4, 5, 2, 3, 4, 5, 6), ncol = 2)
y <- c(1.2, 2.3, 2.8, 4.2, 5.0)
model <- least_squares(X, y)
new_X <- matrix(c(6, 7, 8, 9, 10, 11), ncol = 2)
predictions <- predict_least_squares(model, new_X)
print(predictions)
}
