% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_Predict.R
\name{predict_simple_lm}
\alias{predict_simple_lm}
\title{Predict using Simple Linear Regression}
\usage{
predict_simple_lm(model, new_x)
}
\arguments{
\item{model}{A model object returned by simple_lm.}

\item{new_x}{A numeric vector of new independent variable values.}
}
\value{
A numeric vector of predicted values.
}
\description{
This function predicts new values using a fitted simple linear regression model.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2.2, 2.8, 3.6, 4.5, 5.1)
model <- simple_lm(x, y)
predictions <- predict_simple_lm(model, c(6, 7))
print(predictions)
}
